refactor nil log checking in pollingDeviationChecker
