+++ core/services/fluxmonitor/flux_monitor_internal_test.go
@@ -1,14 +1,10 @@
 package fluxmonitor
 
 import (
-	"testing"
-
 	"github.com/smartcontractkit/chainlink/core/services/eth"
-	"github.com/smartcontractkit/chainlink/core/services/eth/contracts"
 	"github.com/smartcontractkit/chainlink/core/utils"
 
 	"github.com/ethereum/go-ethereum/common"
-	"github.com/stretchr/testify/assert"
 )
 
 func (fm *concreteFluxMonitor) MockLogBroadcaster() *mockLogBroadcaster {
@@ -33,20 +29,4 @@ func (mlb *mockLogBroadcaster) Stop()                                      {}
 
 type MockableLogBroadcaster interface {
 	MockLogBroadcaster() *mockLogBroadcaster
-}
-
-func TestFluxMonitor_PollingDeviationChecker_HandlesNilLogs(t *testing.T) {
-	p := &PollingDeviationChecker{}
-	var logNewRound *contracts.LogNewRound
-	assert.NotPanics(t, func() {
-		p.respondToLog(logNewRound)
-	})
-	var logAnswerUpdated *contracts.LogAnswerUpdated
-	assert.NotPanics(t, func() {
-		p.respondToLog(logAnswerUpdated)
-	})
-	var randomType interface{}
-	assert.NotPanics(t, func() {
-		p.respondToLog(randomType)
-	})
 }
